+++ src/Chirp/Parser.php
@@ -15,37 +15,35 @@ use DecodeLabs\Tagged\Markup;
 
 class Parser
 {
-    public const AT_SIGNS = '[@＠]'; // @ignore-non-ascii
+    protected const UrlPre = '(?:[^-\\/"\':!=a-z0-9_@＠]|^|\\:)'; // @ignore-non-ascii
+    protected const Domain = '(?:[^\\p{P}\\p{Lo}\\s][\\.-](?=[^\\p{P}\\p{Lo}\\s])|[^\\p{P}\\p{Lo}\\s])+\\.[a-z]{2,}(?::[0-9]+)?';
+    protected const Tld = '/\\.(?:com|net|org|gov|edu|uk)$/iu';
+    protected const Path = '(?:(?:\\([a-z0-9!\\*\';:=\\+\\$\\/%#\\[\\]\\-_,~]+\\))|@[a-z0-9!\\*\';:=\\+\\$\\/%#\\[\\]\\-_,~]+\\/|[\\.\\,]?(?:[a-z0-9!\\*\';:=\\+\\$\\/%#\\[\\]\\-_~]|,(?!\s)))';
+    protected const PathEnd = '[a-z0-9=#\\/]';
+    protected const Query = '[a-z0-9!\\*\'\\(\\);:&=\\+\\$\\/%#\\[\\]\\-_\\.,~]';
+    protected const QueryEnd = '[a-z0-9_&=#\\/]';
 
-    public const URL_PRE = '(?:[^-\\/"\':!=a-z0-9_@＠]|^|\\:)'; // @ignore-non-ascii
-    public const URL_DOMAIN = '(?:[^\\p{P}\\p{Lo}\\s][\\.-](?=[^\\p{P}\\p{Lo}\\s])|[^\\p{P}\\p{Lo}\\s])+\\.[a-z]{2,}(?::[0-9]+)?';
-    public const URL_TLD = '/\\.(?:com|net|org|gov|edu|uk)$/iu';
-    public const URL_PATH = '(?:(?:\\([a-z0-9!\\*\';:=\\+\\$\\/%#\\[\\]\\-_,~]+\\))|@[a-z0-9!\\*\';:=\\+\\$\\/%#\\[\\]\\-_,~]+\\/|[\\.\\,]?(?:[a-z0-9!\\*\';:=\\+\\$\\/%#\\[\\]\\-_~]|,(?!\s)))';
-    public const URL_PATH_END = '[a-z0-9=#\\/]';
-    public const URL_QUERY = '[a-z0-9!\\*\'\\(\\);:&=\\+\\$\\/%#\\[\\]\\-_\\.,~]';
-    public const URL_QUERY_END = '[a-z0-9_&=#\\/]';
-
-    public const URL_ALL = '/(?:' .
-        '(' . self::URL_PRE . ')' .
+    protected const Url = '/(?:' .
+        '(' . self::UrlPre . ')' .
         '(' .
         '((?:https?:\\/\\/|www\\.)?)' .
-        '(' . self::URL_DOMAIN . ')' .
-        '(\\/' . self::URL_PATH . '*' .
-        self::URL_PATH_END . '?)?' .
-        '(\\?' . self::URL_QUERY . '*' .
-        self::URL_QUERY_END . ')?' .
+        '(' . self::Domain . ')' .
+        '(\\/' . self::Path . '*' .
+        self::PathEnd . '?)?' .
+        '(\\?' . self::Query . '*' .
+        self::QueryEnd . ')?' .
         ')' .
         ')/iux';
 
-    public const USERNAME_LIST = '/([^a-z0-9_\/]|^|RT:?)([@＠]+)([a-z0-9_]{1,20})(\/[a-z][-_a-z0-9\x80-\xFF]{0,24})?([@＠\xC0-\xD6\xD8-\xF6\xF8-\xFF]?)/iu'; // @ignore-non-ascii
-    public const USERNAME_MENTION = '/(^|[^a-z0-9_])[@＠]([a-z0-9_]{1,20})([@＠\xC0-\xD6\xD8-\xF6\xF8-\xFF]?)/iu'; // @ignore-non-ascii
-    public const USERNAME_REPLY = '/^(' . self::WHITESPACE . ')*[@＠]([a-zA-Z0-9_]{1,20})/'; // @ignore-non-ascii
+    protected const UserList = '/([^a-z0-9_\/]|^|RT:?)([@＠]+)([a-z0-9_]{1,20})(\/[a-z][-_a-z0-9\x80-\xFF]{0,24})?([@＠\xC0-\xD6\xD8-\xF6\xF8-\xFF]?)/iu'; // @ignore-non-ascii
+    protected const Mention = '/(^|[^a-z0-9_])[@＠]([a-z0-9_]{1,20})([@＠\xC0-\xD6\xD8-\xF6\xF8-\xFF]?)/iu'; // @ignore-non-ascii
+    protected const Reply = '/^(' . self::WhiteSpace . ')*[@＠]([a-zA-Z0-9_]{1,20})/'; // @ignore-non-ascii
 
-    public const HASHTAG = '/(^|[^0-9A-Z&\/\?]+)([#＃]+)([0-9A-Z_]*[A-Z_]+[a-z0-9_üÀ-ÖØ-öø-ÿ]*)/iu'; // @ignore-non-ascii
-    public const WHITESPACE = '[\x09-\x0D\x20\x85\xA0]|\xe1\x9a\x80|\xe1\xa0\x8e|\xe2\x80[\x80-\x8a,\xa8,\xa9,\xaf\xdf]|\xe3\x80\x80';
+    protected const HashTag = '/(^|[^0-9A-Z&\/\?]+)([#＃]+)([0-9A-Z_]*[A-Z_]+[a-z0-9_üÀ-ÖØ-öø-ÿ]*)/iu'; // @ignore-non-ascii
+    protected const WhiteSpace = '[\x09-\x0D\x20\x85\xA0]|\xe1\x9a\x80|\xe1\xa0\x8e|\xe2\x80[\x80-\x8a,\xa8,\xa9,\xaf\xdf]|\xe3\x80\x80';
 
-    public const BASE_URL = 'https://twitter.com/';
-    public const SEARCH_PATH = 'search?q=%23';
+    protected const BaseUrl = 'https://x.com/';
+    protected const SearchPath = 'search?q=%23';
 
     /**
      * Convert plaintext tweet to HTML
@@ -62,7 +60,7 @@ class Parser
         // Urls
         $text = $this->processUrls($text);
 
-        // Hashtags
+        // HashTags
         $text = $this->processHashTags($text);
 
         // Usernames
@@ -86,15 +84,21 @@ class Parser
     protected function processUrls(
         string $text
     ): string {
-        return preg_replace_callback(self::URL_ALL, function ($matches) {
+        return preg_replace_callback(self::Url, function ($matches) {
             list($all, $before, $url, $protocol, $domain) = array_pad($matches, 7, '');
             $url = $this->esc($url);
 
-            if (!$protocol && !preg_match(self::URL_TLD, $domain)) {
+            if (!$protocol && !preg_match(self::Tld, $domain)) {
                 return $all;
             }
 
-            $href = (!$protocol || strtolower($protocol) === 'www.' ? 'https://' . $url : $url);
+            $href = (
+                !$protocol ||
+                strtolower($protocol) === 'www.'
+            ) ?
+                'https://' . $url :
+                $url;
+
             return $before . $this->wrapUrl($href, 'url', $url);
         }, $text) ?? $text;
     }
@@ -106,10 +110,10 @@ class Parser
     protected function processHashTags(
         string $text
     ): string {
-        return preg_replace_callback(self::HASHTAG, function ($matches) {
+        return preg_replace_callback(self::HashTag, function ($matches) {
             $replacement = $matches[1];
             $element = $matches[2] . $matches[3];
-            $url = self::BASE_URL . self::SEARCH_PATH . $matches[3];
+            $url = self::BaseUrl . self::SearchPath . $matches[3];
             return $replacement . $this->wrapUrl($url, 'hashtag', $element);
         }, $text) ?? $text;
     }
@@ -121,7 +125,7 @@ class Parser
     protected function processUsernames(
         string $text
     ): string {
-        return preg_replace_callback(self::USERNAME_LIST, function ($matches) {
+        return preg_replace_callback(self::UserList, function ($matches) {
             /**
              * @var string $before
              * @var string $after
@@ -135,12 +139,12 @@ class Parser
             if (!empty($listname)) {
                 $element = $username . substr($listname, 0, 26);
                 $class = 'list';
-                $url = self::BASE_URL . $element;
+                $url = self::BaseUrl . $element;
                 $suffix = substr($listname, 26);
             } else {
                 $element = $username;
                 $class = 'user';
-                $url = self::BASE_URL . $element;
+                $url = self::BaseUrl . $element;
                 $suffix = '';
             }